Trilinos FROSch Solver interface
This commit introduces a solver interface to the thirdparty FROSch
solver, which is a algebraic parallel preconditionier that works
with simple Poisson type systems as well as Stokes typed saddlepoint
systems.

It is highly recommended to use a partitioner like parmetis to create
the mesh partition used by FROSch, since the solver relies on simply
connected subdomains.

Besides test applications, an interface to the gendie solver was
implemented.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>