use cuda_device_synchronize after each cuda kernel call
The unified memory runtime does trigger a memory copy if data does not
reside in the current memory. Unfortunately this memory copy is
triggered regardless of any running cuda kernel or cpu thread.
Thus the data received may not be the final result of an ongoing
operation.
To circumnavigate this issue, this commit de facto introduces a new
paradigm, that any operation/kernel/thread must be finished if the
program counter returns to the next operation.
The only exception from this rule are operations that are labeled as
asynchronous like non-blocking MPI or std::future handles.