added 3D flow-around-a-sphere-inside-a-cylinder benchmarks (bench7+8)
This commit updates both the "simple" CCND as well as the "DFG95" CCND
applications to support a new pair of "DFG95-style" benchmarks, which
simulate a 3D flow around a sphere inside a cylinder. The domain can
be seen as an alternative extension of the 2D case into 3D, where the
domain is turned into a cylinder rather than a rectoid and the obstacle
is turned into a sphere rather than being extruded into a cylinder.

The first experiments have shown that choosing a maximum inflow velocity
of 1.0 leads to a stable steady-state solution -- this configuration has
been chosen as the benchmark case for the steady-state solvers and has
been named "bench 7".

For a maximum inflow velocity around 2.0, the solution still seems to
converge to a steady state, however, the current configuration of the
steady CCND solvers fail to solve these problems, however, the unsteady
CCND versions converge.

For a maximum inflow velocity of 4.0, the flow is unsteady and it also
reaches a perdiodic flow after a few seconds of simulation time (not to
be confused with "runtime", which may be several days/weeks). In this
scenario, the periodic flow is actually pretty interesting, because a
single period has two minima and two maxima of drag and lift of different
amplitudes. This configuration has been chosen as the benchmark case for
the unsteady solvers and has been named "bench 8".

This commit also adds an appropriate coarse mesh file with 256 hexahedral
cells named "flowbench_p3d_00_hexa_256.xml".

Furthermore, this commit also adds the Analytic::Common::PoiseuillePipeFlow
function, which is used as an inflow boundary conditions for these new
benchmark problems.

This commit also enables the iso-parametric transformation as the default
case for the "simple" CCND applications.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>