ported dbg-amavanka debug app for CUDA unified memory
Well, the new unified memory interface makes the AmaVanka smoother simple to
use on CUDA devices, however, up to the current version of CUDA, the B[C]SR
algorithms implemented in cuSPARSE only support square blocks. In consequence,
in each matrix-vector product with a Stokes-like saddle-point matrix, the
products with the submatrix A are performed on the GPU, whereas the products
with the submatrices B and D are performed by the CPU, which in consequence
triggers several up/downloads of vector data per matrix-vector product, making
the CUDA version of the AmaVanka (and the normal matrix-vector product) slow
as hell.

The only viable solution to this is to find a BCSR matrix-vector CUDA kernel
(or write one by ourselfes) which works for non-square blocks as well to avoid
the constant up- and downloading of data during each matrix-vector product.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>