Lower convergence limits for some cuda solver tests
In addition deactivated all cuda ilu tests and some ugly unstable
solvers for cuda.
The cuda ilu implementation has no forseable future.
The ilu decomposition methods are deprecated with no perspective on
replacement. The triangular solver methods are deprecated and there
replacement needs to now the rhs vector already in init_numeric (in the
analysis and get_buffer_size methods). This may be not compatible with
our "the preconditioner does not know the vector apriori" concept.