disable SolverExpression collection in Statistics by default
This commit disables the collection of SolverExpressions, which are accumulated
by Statistics::add_solver_expression, unless the user explicitly enables this
by setting Statistics::enable_solver_expressions to true.

The reason for this is that the Statistics class could blow any memory boundary
by accumulating ridiculous amounts of solver expressions in applications that
perform millions of time-steps during a single simulation run.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>