added StructUnitCubeFactory for ConformalMesh and StructuredMesh types
This new class creates a structured unit cube mesh with lexicographic numbering
of the vertices and elements. In addition that, this commit also introduces two
new methods to create a PartiDomainControl: first, an overload of "create" that
accepts a base-mesh node as a parameter as well as a "create_rectilinear"
function, which utilizes the StructUnitCubeFactory to create a rectilinear
base mesh for the domain control.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>