added 2D-Q2[/P1dc] Poisson and Stokes multigrid performance benchmark apps
This commit introduces two new parallel benchmarking applications for the 2D-Q2
Poisson and 2D-Q2/P1dc Stokes problems on the unit square domain.
These benchmarks are similar to the 'poisson_bench_mg' application in the sense
that they collect and output a lot of statistics, but they differ in the fact
that they do not read in a mesh file and refinement level, but create the mesh
hierarchy on the fly based on a maximum memory usage request supplied by the
user, which is more appropriate for system benchmarking.

The Poisson benchmark uses a PCG-Multigrid-Jacobi solver, whereas the Stokes
benchmark uses a Richardson-Multigrid-AmaVanka solver. The multigrid solvers
for both benchmarks do not use any "real" coarse grid solver, but only use the
smoother on the coarse mesh, which is (usually) a 1x1, 2x2 or 3x3 mesh anyways.

These two benchmark apps have also been added to the regression test system.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>