added support for ignored NaNs in UnitFilterBlocked to emulate slip filter
This commit extends the functionality of the UnitFilterBlocked class to enable
the emulation of an axis-aligned SlipFilter by explicitly disabling the filtering
of individual components on a per-DOF basis by settings the not-to-be-filtered
component values to NaN (not-a-number). Please note that this functionality is
disabled by default to avoid unexpected (false-positive) behavior when a filter
contains NaNs values due to a bug or runtime error.

Example: If you want to emulate a 3D SlipFilter that filters the Y-component
then set the values of the filter to (NaN, 0, NaN). Note that you can generate
a NaN for any data type by using the Math::nan<DT_>() function.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>