added a bunch of mesh files for FBM method testing
This commit adds three groups of mesh files:

* flowbench_c2d_05_fbm_aligned_quad_76.xml
* flowbench_c2d_06_fbm_quad_20.xml
These two meshes are FBM meshes for the 2D "flow-around-a-cylinder" benchmark,
where one mesh is aligned with the FBM interface and the other one is not.

* unit_ring_quad_32.xml
* unit_ring_fbm_quad_4.xml
* unit_ring_fbm_aligned_quad_32.xml
These three meshes are used to discretize the unit-ring domain with radius
[0.5, 1.0] around the origin (0,0). The first mesh is a "real" boundary mesh,
i.e. an actual discretization of the ring domain. The second mesh is a simple
2x2 square mesh representing the domain [-1.25,+1.25]^2 thus enclosing the
unit ring domain and the third mesh is also a mesh representing the domain
[-1.25,+1.25]^2 but with mesh elements aligned to the unit ring domain.

* ypipe_2d_quad_24.xml
* ypipe_2d_fbm_quad_6.xml
These two meshes represent a symmetric 2D Y-pipe domain with one inflow on the
left side and two outflows at the right side, where the outer boundaries are
described by Bezier splines. The first mesh is a "real" boundary mesh
representing the actual Y-pipe domain, whereas the second mesh is a simple
unaligned 3x2 quad mesh for FBM.

Additionally, this commit also adds the 'ypipe_2d_boundary.vtp' file, which
is can be opened in ParaView to visualize the Y-pipe domain for post-processing.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>