implemented UnitFilter[Blocked]::filter_weak_matrix_rows() functions
The scalar and blocked unit filters have been extended by a functions which
allows to replace a set of rows of a matrix A by the rows of another matrix M,
scaled by the values stored in the filter.

These functions replace all rows of the system matrix A, whose row index is
included in this filter's indices, by the corresponding rows of a given donor
matrix M, which is typically a mass matrix, scaled by the values which are
stored in this filter.

This functionality can be used to implement a weak form of Dirichlet boundary
conditions, which is also used to employ mass-conserving fictitious boundary
conditions in a Stokes system.

Corresponding unit-test for the new functions have also been added.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>