added new family of CCnD applications including FBM support
This commit introduces a set of basic steady and unsteady CCnD applications,
which are derived from the "dfg95_bench" CCnD applications. These new CCnD apps
have been split into a few major classes to drastically reduce the amount of
"dumb-copy-&-paste" that was involved in the DFG95 benchmark applications,
therefore drastically reducing the amount of code in the actual application
source files. This is still a good way from the "sophisticated CCnD" that I
have been working (somewhat unsuccessfully) on for several years now, but I
hope that this approach will be more useful and easier to extend than all
previous approaches.

The "ccnd_common.hpp" header defines some helper classes as well as the base
classes "CCND::DomainLevelBase" as well as "CCND:SystemLevelBase", which
can be used by applications as the "DomainLevel" and "SystemLevel" classes or
as base-classes for custom implementations thereof.

The "ccnd_steady_base.hpp" defines the "CCND::SteadyAppBase" class, which
is used by the steady-state applications as a base-class for their actual
Application class. Pretty much the entire steady-state CCND code is outsourced
into this base-class with the exception of the filter assembly, the RHS assembly
and the post-processing of the solution (error computation, drag&lift forces).

The "ccnd_unsteady_base.hpp" defines the "CCND:UnsteadyAppBase" class, which
derives from the "CCND::SteadyAppBase" class and extends its functionality by
implementing a BDF(2) time stepping scheme for the time discretization. Again,
this class is used by applications as a base-class for their Application class
to outsource most of the actual CCnD code.

There is one technical stunt that had to be pulled off to simplify the code of
the [Uns|S]teadyAppBase classes drastically: One has to define both the
"CCND:DomainLevel" and "CCND::SystemLevel" classes *BEFORE* including the
"ccnd_steady_base.hpp" or "ccnd_unsteady_base.hpp" header files. So the first
few lines of a steady-state CCND application may look like this:

<code>

#include "ccnd_common.hpp"

namespace CCND
{
  // either typedef or custom "DomainLevel" class
  typedef DomainLevelBase DomainLevel;

  // either typedef or custom "SystemLevel" class
  typedef SystemLevelBase SystemLevel;
}

#include "ccnd_steady_base.hpp"

namespace CCND
{
  class Application :
    public CCND::SteadyAppBase
  {
    // actual application class follows...

</code>

One important addition in comparison to the old DFG95 benchmark apps is that
the new CCnD applications now also support fictitious boundary conditions.
It turned out that by using the new StokesFBMAssembler class, the FBM could
be directly integrated into the basic CCND applications, because only a few
additional lines of code had to be included for that and the overall program
structure did not change drastically enough to justify an extra app for that.

Currently, there exist 3 applications in the new CCnD framework:

* ccnd-steady-dfg95-2d/3d
The good old 2D/3D steady-state flow-around-a-cylinder benchmark (aka "bench1")
including all the fancy post-processing like line- and volume-based drag & lift
force calculation, etc.

* ccnd-unsteady-dfg95-2d/3d
The good old 2D/3D unsteady flow-around-a-cylinder benchmarks with steady and
time-dependent inflow conditions a.k.a. "bench2" and "bench3", respectively.
Also includes the post-processing tools for this benchmark, including the two
Python scripts "bench2plots.py" and "bench3plots.py" to analyze logs and
generate fancy drag-&-lift curves for your publications. You're welcome :)

* ccnd-steady-taygre-2d
The 2D steady-state Taylor-Green Vortex benchmark on the unit-square, which
primarily exists to analyze and verify proper convergence orders.

Two final notes:

Firstly, the new CCnD application framework code is far from "finished"; this
code will be extended and improved continuously in the future. Obviously, more
applications based on this framework will be added in the future.

Secondly, the "old" DFG95 benchmark apps in "applications/dfg95_bench" have
become obsolete with this commit, however, they will not be removed as for now
because there are some developers working on derived versions of these.
However, it is recommended to derive new CCND applications based on the new
framework from now on, because the old DFG95 bench apps will be removed at some
point.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>