various improvements of new CCND and added 2D Y-pipe benchmark CCND
This commit adds a bunch of improvements and experimental features to the new
CCND framework.

Firstly and foremostly this commit includes a hotfix for the nonlinear defect
computation, which was not entirely correct in the case of FBM. The results do
not seem to differ for the benchmarks we have done so far, but they may differ
in more turbulent cases. We still need to evaluate whether this is the proper
way of doing things in the FBM case anyways.

Secondly, this commit outsources some of the FBM assembly into the base classes
thus putting the actual application codes on a more aggressive diet.

Thirdly, the solver infrastructure has been improved in the sense that no
multigrid hierarchy is created if only a single level is desired. This also
improves the handling of the case where only the DirectStokesSolver is used.

Fourthly, the FBM velocity and pressure masks are now exported to the VTK files.

Fifthly, a new CCND for the 2D Y-pipe benchmark has been added, which also
supports FBM simulations. This benchmark is not wildly interesting on its own,
but it will probably serve well as a simple testing ground for FBM simulations.
Right now, the Y-pipe domain has equal length outflows, but we may add meshes
with different length outflows in the future as this is known to be an
interesting test case for various boundary condition formulations.

Finally, the Geometry::ChartHitTestFactory class has been added, which can be
used to created a Geometry::MeshPart from a Geometry::Atlas::Chart, which is
especially useful in the FBM case.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>