Added CGALSurfaceMesh to charts
This commit introduces the possibility to parse an off file as
surface chart as part of the FEAT file format.
See the FEAT fileformat documentation page for details.

Also added a bit of necessary CGAL functionality.

Furthermore added a cgal chart to the simple scalexa mesh.

TODO: Allow to add a CGALWrapper as chart directly to a MeshAtlas.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>