Make feat3 cuda 12 compileable and adjust tests
This commit fixes the cuda ilu solver to remove deprecated
cusparse calls for solving triangular systems.
The changes are introduced based on the used cusparse version
and do not change anything in case you use a cuda version lower
12.

This change does NOT remove deprecated calls to the cusparse ilu
calls, which will lead to incompatibility to cuda >= 13 releases.

Furthermore adjusted the meta_matrix-apply-test.
It appears that uneven matrix sizes lead to a Warp Misaligned Address
error in the cudaFree call.
Further research is required where exactly the alignment breaks...

Also activated a ilu solver-test.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>