Removed cgal chart parser and added static factory function
Due to consistency errors regarding filepath, it was decided against
the possibility to define a CGALMeshSurface directly inside
the FEAT mesh file format.

A new static factory function was implemented to directly create a unique ptr
of a CGALSurfaceMesh object, streamlining the creation of a new chart object.

Also adpated the documentation to at least hint to the existence of a
cgal based surface mesh.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>