Fixed unit test system
This commit fixes a bug in the CMakeLists by swapping the main
test identifier with the generic test identifier, which now allows
an internal filter to correctly start the defined unit tests.
Also removed none tests for the simple reason that there is no need
anymore since every test is "generic" by default.

This has the added benefit, that in the case of only generic tests,
ctest can now be run threaded.

Also fixed a few small errors connected to the test system  on the way.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>