Refactored ctest system
Adjusted test system so that, by default, we always register tests
running on all backends (if they are enabled), which is referred to
as <test>_all .
Since this conflicts somewhat with additional memtests for example
valgrind or cuda_memtest, which previously we directly included in the
standard tests, these test are now additinoaly (and explicitly) defined
as <test>_valgrind resp. <test>_cudamemcheck and are only run
on backends generic resp. cuda.

This has the added benefit, that we now can simply run memtest
with ctest -R valgrind .

Additionally mkl backend was added to the valgrind tests in the lafem
library.

Warning: CGAL + valgrind appears to be broken.
See https://github.com/CGAL/cgal/issues/3094

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>