disabled LAFEM serialization tests for __float128 data type
These tests have been disabled because the current implementation of the
serialization of the LAFEM containers always converts the data arrays to double,
which results in loss of precision in case of quad precision. This is not a
regression, but the serialization has to be redesigned form scratch to fix this
issue.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>