added Geometry::MaskedBoundaryFactory class
This factory class is an extension of the Geometry::BoundaryFactory class, which
allows to create a boundary meshpart for only a part of the entire mesh boundary,
which is required to create the proper boundary meshparts for partitioned meshes.

The old Geometry::BoundaryFactory class always creates a single mesh part for
the entire boundary of a (local) mesh. In consequence, this class cannot be used
in the partitioned case, because the class would create a mesh-part for the
entire boundary of the patch local to each process instead of the local patch
portion of the global boundary.

The new Geometry::MaskedBoundaryFactory allows to "mask out" a set of facets
which should not be treated as boundary facets even if they belong to the
local patch boundary. In the partitioned case, one needs to add all halo
mesh-parts, which represent the inter-patch communication regions, to the
factory object. Then, when the factory is used to create the boundary mesh-part,
the facets that belong to any of the halos, will not be treated as boundary
facets and therefore the mesh-part will represent the local portion of the
global boundary, which may legitimately be an empty mesh part if the patch is
an inner patch of the partitioning.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>