added Geometry::RootMeshNode::extract_patch for explicit element list
This new overload of the extract_patch() function for the RootMeshNode allows
to extract a patch from a RootMeshNode by only supplying a list of element
indicies which should be included in the extracted patch. This function also
splits the mesh-parts, halos and patch mesh-parts of the root mesh node, if
desired.

This function adds a new patch for the rank -1 to the mesh node that the patch
has been extracted from.

This new overload will be used by VoxelDomainControl to "deslag" voxel meshes.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>