added Control::Domain::VoxelDomainControl class
A lof of the multi-layered recursive partitioning and management code from the
Control::Domain::PartiDomainControl class has been outsourced into a separate
base class named Control::Domain::PartiDomainControlBase, which can now also
be used by other sophisticated domain control classes.

The new Control::Domain::VoxelDomainControl class provides a domain control
which always starts off at a MxN (2D) or MxNxL (3D) rectilinear/rectoid mesh,
which discretizes the bounding box of the actual computational domain, and
during the refinement and partitioning process the domain control "drops" all
elements which do not intersect with the computational domain, therefore
leading to a "Voxel" approximation of the computational domain whose resolution
improves with each refinement level. This domain control supports recursive
partitioning and therefore multi-layered multigrid hierarchies.

The voxel approximation is created by integrating a so-called "slag mask", which
is effectively a monochrome 2D/3D "bitmap" or "voxmap". This slag mask has to
be supplied by the user and is typically generated by using appropriate helper
classes and functions from either a chart object, which may be provided in the
form of a FEAT mesh file, or from an OFF file containing a surface triangulation,
which is then processed with the help of the CGAL library. Alternatively, a
slag mask can also be created from a simple lambda expression, but this is
primarily only useful for testing and debugging.

Currently, there is no way to read in a slag mask directly from a file, but this
feature will definitely be added in the future once the development team has
agreed on an approprimate file format for the voxmap.

One downside of this voxel domain approach is that the mesh hierarchy is now
"non-conforming" by construction in the sense that a coarse mesh and a refined
mesh do not represent the same discrete computational domain anymore, because a
subset of the refined elements gets dropped after each refinement step. This in
turn leads to a "non-conforming" multigrid hierarchy, which can (and usually
does) lead to problems with convergence. This is not a bug, but a mathematical
drawback of the chosen approach -- we all have to die somehow...

The usage of the VoxelDomainControl is mostly compatible to the usage of the
other domain control classes. One major exception is the assembly of grid
transfer operators, which now has to take the aforementioned non-conformity of
the refinement into account. To make transition as simple as possible, a new
helper class Control::VoxelTransferAssembler has been added, which takes care
of assembling the proper grid transfer operators for scalar, blocked or stokes
systems by using the appropriate assembly function.

This commit also adds two simple sample applications "poisson_voxel_mg" and
"stokes_voxel_ypipe", which demonstrate the usage of the VoxelDomainControl
class as well as the VoxelTransferAssembler helper class.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>