implemented SymbolicAssembler::assemble_graph/matrix_intermesh() functions
These new functions can assemble a graph/matrix structure with trial and test
spaces defined on different meshes, however, the caller has to supply an
adjactor (e.g. and Adjacency::Graph), which describes the element relation
between the test and trial spaces.

This is mostly a mildy refactored version of assemble_graph/matrix_2lvl, which
already performed this exact job with the 2-level refinement adjactor that is
already defined in Geometry::Intern::CoarseFineCellMapping. In consequence,
the 2lvl assembly now simply calls the intermesh assembly and passes the
aforementioned coarse-2-fine-cell-mapping as the intermesh-mapping.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>