Added mpiexec to the ctest run calls
This commits prepends to all sequential test calls an additional
variable FEAT_CTEST_RUNNER which is (for now) empty if we have compiled
without MPI and contains a one process mpiexec call if FEAT is compiled
with MPI.
This arrives from an oversight of the new test calls in conjuncture with
slurm enviroment variables.

Attention: This also means that the memcheck tests are wrapped within
an mpiexec call. If the executable, for example valgrind, is not
correctly configured to run within an mpicall, this will obviously lead
to false negative test results.
Since the combination of mpi + memcheck are not used in the regression
tests by default, this should be an acceptable way.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>