implemented surface-based drag/lift/side forces computation for DFG95 bench
This commit updates the surface-based (previously called 'line' variant) computation
of the drag & lift coefficients for the 2D/3D DFG95 benchmarks in the ccnd application.
The previous implementation was based directly on the formulae given in the original
DFG95 paper by Schaefer & Turek, whereas the new version is based on the stress
formulation as speicifed in the paper from Giles et al:

Adaptive error control for finite element approximations of the lift and drag coefficients in viscous flow

Note: This is also the paper that proved that the "volume-based" variant of the
drag & lift computation, which uses the final (unfiltered) defect vector to compute
the body forces, can actually achieve an order of O(h^4) for Q2/P1dc in optimal circumstances.

The new surface-based method has been implemented to support the correct body forces
compuation even for real 3D meshes, e.g. the "flow-around-a-sphere" benchmark, which
was not possible with the old line-based formulation, since it has explicitly been
derived only for 2.5D meshes in 3D.

The coefficients computed by the new surface-based method differ slightly from the
ones computed by the old line-based method by a factor of up to 2, which shouldn't
be a big issue though, since we usually prefer to analyze the volume-based coefficients
anyways due to the higher order.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>