Print out visible cuda devices per rank in poinsson mg benchmark
As of now, it is highly erratic which GPU is visible on which rank on
different cluster environments.
From one device per rank, over all devices in a numa/socket domain or
all allocated devices to all devices on the node, even those not
allocated, all possabilities have been observed.
Thus it is very helpfull to have the information about the visible
devices on hand for each rank in the very execution of the benchmark, as
the visible devices can differ by the way of execution, too.

Common pitfalls are
* manually setting CUDA_VISIBLE_DEVICES at all or setting it in the wrong environment.
* using mpirun instead of srun, where srun would manage the visible devices in accordance with the allocated sockets
* forgetting to update the --gres::GPU::x number of GPUs per node according to the number of tasks/ranks per node
* loosing track of the rank to node mapping, e.g. round robin or linear mapping