Cuda include ready header files
This commit changes a few chosen headers, namely lagrange2/details.hpp,
tiny_alegbra.hpp and runtime.hpp, as well as assertion.hpp.
For this, a few macros and internals had to be adapted.

Also provided a bit of abstraction for the trafo and assembly operations to allow
the actual kernels to be repurposed in other functions.
These abstractions were also made cuda aware.

Furthermore fixed a few bugs regarding these changes and clang.

Also provided a few useful cuda math function interfaces, as well as
a bit additional interfaces for the cuda runtime api.

Since this change allows for the nvcc compiler to potentially be
aware of the complete FEAT library, this changes could lead to
unforseen bugs in the future, which will be handled if they arise.

Also, not tested with the ipx family of compilers, yet.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>