Voxel assembly classes
This commit introduces a new branch of the kernel, the voxel assembly
module. Besides a large amount of adapter code, the main contribution
are the Voxel<Operator>Assembler classes which provide a familiar
interface for assembly of finitie element matrices.
The main difference to the present assembly routines is the dependency
on a given cell wise coloring, which pretty much requires a at least
substructered mesh for optimal performance.
In exchange, the assembly routines run fully threadparallel, allowing
the use of an efficient cuda backend, as well openOMP based routines.
The used backend is, as already established, chosen
through the PrefferedBackend parameter and *should* allow for an optimal
use of compute resources, without specializing appllications.

For now, only Q2 FE space with standard trafo are supported, more to
come. No need for structered meshes (for now).
Also it is of course planed (and possible) to specialize kernel routines
for the case of actual voxel meshes.

Also added tests.

TODO: Optimize the burgers typed assembler by using warp based
threading strategies.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>