Rule of five for CGALWrapper and threadsafe check
This commit introduces the rule of five for the CGAL wrapper
by defining a move ctor as well as an move assignment, while
deleting the copy ctor and assigment operators.

Additionally a check was introduced to guarantee thread safe operations
in CGAL, allowing to use omp routines around the cgal wrapper functions
without creating a CGALWrapper object for each thread.
Through this change, CGAL can only be included if either BOOST is
compiled with threading capabilities or openMP is included.

Additionally a small parallel threading test was introduced for the
point inside method.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>