added Geometry::VoxelMap class and corresponding voxel-map-gen tool
This commit introduces the Geometry::VoxelMap class, which is now used by the
VoxelDomainControl to manage the internal voxel map, which was previously called
'slag mask'. In addition to that, a new tool 'voxel-map-gen' has been added
which can be used to create a voxel map file from an OFF model file by utilising
the CGAL third-party library, so that the actual voxel map generation can be
outsourced from the domain control setup, since the voxel map generation can be
a tremendously time consuming process. Voxel map files are stored in a custom
binary file format, which supports ZLIB-based compression of the actual map.

This change was necessary because in the case of very fine geometries, like the
"Wendelverteiler" domains, the previous "slag mask" resolution was equal to the
finest mesh width, which may have been too low to represent all the geometric
details. Now, the voxel map resolution can be chosen independent of the fine
mesh resolution.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>