Refactoring interfaces
This commit changes a few class names, as well as a few cuda specific
calls.

Furthemore renamed the CUDA_FUNC macro to CUDA_HOST_AND_DEVICE.
This macro, as well as the CUDA_HOST and CUDA_DEVICE macro
are intended for functions and methods defined in header files
both included by cuda and "normal" cpu compile units.

If not compiled by nvcc, the macro is simply empty.
Else it behaves as the nvcc specific __host__ and __device__ macros.
If you do not know what they do, you can simply ignore the flags,
as not providing anything is the same as providing CUDA_HOST,
which is normally what you want.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>