Fix container write out and add mtx out buffer
This commits fixes a few things in the write out methods of the
lafem containers and adds a fixed sizes heap io buffer for all
MTX format write outs, which should reduce the time to write out
significantly.

This approach neither provides significant performance boosts
read in operations and binary write outs, so for this cases
the internal buffer is not overwritten.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>