addes stringify_bytes function to format memory usage
The output unit is chosen based in the input value bytes in a way, so that there are at
most 3 leading digits and for any size >= 1000 bytes, and the output will contain as many digits
as were specified by the precision parameter and it will be padded to the right to as many
characters as specified by the width parameter excluding the unit postfix.
For example, the well-known HD 3.5" floppy disk has a size of 1457664 bytes, which results in
an output of "  1.390 MiB", and a standard CD-ROM has a size of 681574400 bytes, which results
in an output of "650.000 MiB".

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>