Corner Singularity function for Laplace operator
This commit introduces two version of the CornerSingularity
function, representing the lowest frequency of the singularity
found in re-entry corners for the 2D Laplace equation.

While the standard version is simply an analytic function,
providing values, gradients and hessians,
for testing purposes also an implementation based on
Polar-Coordinates is introduced.

Since there is no wrapper for transforming Polar-Coordinate based
functions into the euclidean space yet, a new wrapper providing
this functionality is also introduced for the case of scalar functions.

Also added a unittest.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>