Use previus improvements for AlPiNe iteration
Change the estimated tolerance for the linear solver in the alpine case
to use not only the defect improvement of the previous step but also
from the step before it (if applicable).
Furthermore use linear resp. quadratic estimate depending on the actual
nl step we are doing.

TODO: The estimate probably is not optimal yet, but helps with stabilizing
the solver in case of very hard non-linear problems.
We should think about this again...

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>