added ---build-info, ---debug-break and ---print-pid to Runtime::initialize()
These triple-hyphened command line options are parsed and interpreted directly
by the Runtime class. To ensure compatibility with the SimpleArgParser class,
the query_unsupported function of the SimpleArgParser has been adjusted to
silently ignore/skip/accept all options beginning with a triple-hyphen.

The ---build-info option makes the runtime write various informations about the
current build configuration to std::cout, including the git hash, build-id
and whether various FEAT_HAVE_*** defines are set. In an MPI run, only rank 0
plots the build information.

The ---debug-break [<ranks...>] option is only available when compiling with
the Visual Studio compiler under Windows and it triggers a breakpoint for the
specified MPI ranks, thus enabling the user to attach a debugger directly at
program startup. This is necessary since it is not possible to start an MPI
job directly in the VS debugger, so one has to attach to a running process.

The ---print-pid option is also only available on Windows and it will print
the Windows process ID for each MPI rank onto std::cout.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>