implemented GlobalMaskedBoundaryFactory class
As it turns out, using the MaskedBoundaryFactory class for generating the
boundary meshpart in the case of partitioned meshes may produce incorrect
results in some cases.

A minimal example is the following: consider the L-shape domain partitioned into
3 patches with 1 element each:

   +---+
   | 2 |
   +---X---+
   | 0 | 1 |
   +---+---+

On ranks 1 and 2, the vertex 'X' is correctly considered to be a boundary vertex,
because it is adjacent to a boundary edge, however, on rank 0 this vertex is only
adjacent to halo edges, but not to any boundary edge, thus it would be incorrectly
treated as an inner vertex.

The new GlobalMaskedBoundaryFactory class solves this problem by explicitly
synchronizing the detected boundary entities via the supplied halos.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>