Allow for external include for specfic packages
This commit allows, by providing the new --prefer_external_libraries
configure flag, to use the findPackage functionality of CMake to link
against certain packages (for now Boost and CGAL), removing the need
to download the libraries explicitly, which is consequently disabled
for the allegeable libraries.
Naturally you have to provide the include pathes to the
<package>configure.cmake files yourself, for example by extending your
PATH variable with the lib path of the install directory.
For details, refer to the findPackage documentation of CMake.
In the best case, the correct pathes are already set by the system
for you.

For now, you have to provide external packages for all thirdparties
allegeable for external linkage which you want to use.
This behavior is planned to be refactored sooner or later, but requires
major refactoring in the way we link against libraries.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>