Add Frankes function as analytic function
Added a 2D scalar function with a saddlepoint structure on the unit quad
where the peaks are 2D Gaussian bell curves.
This function is also known as Frankes function and is a well known
test problem for interpolation and optimization algorithms.

Also added a 3D version, which is simply a composite of Frankes function
and a Gaussian bell curve in z-direction.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>